Added support for GitHub/BitBucket sub-repos in 'packages.json' files
* added optional use of GitHub or BitBucket single package providers as the source URL for individual packages within a generic package repository (packages.json)
* additionally, when using a GitHub or BitBucket package provider, 'version' becomes an optional parameter

RATIONALE: This modification adds an additional level of indirection which, if used while avoiding the optional 'version', allows
for the creation of a more general 'packages.json' description file. This more generic 'packages.json' file would only need to be
modified to add or remove package references; otherwise, it can remain unchanged while continuing to develope the individual packages.