Better handle random github (or other repo) errors
A recent fileserver disruption at Github caused a few random package
repos to become unavailable. Part of the way they failed was to return
an HTTP status of 0. The httplib.HTTPException raised is not among (or
a parent of) the exceptions currently being caught, so this completely
broke attempts to install/upgrade packages.