Fix executing loaders as dependencies are installed
Previously, if more than one dependency was installed in ST3, the
installation of anything but the first would fail with an
ImportError. This appears to be due to the zip import mechanism
being used by sublime_plugin caching the file listing of the
.sublime-package file the first time a package submodule was
imported.

This fix manually uses the zipimport module, and clears out the
zip file directory cache before trying to import the loader for a
new dependency.