Don't re-enable package when removal is deferred
ST2 on Windows does not release file handles to .py plugin files if it
encounters an Exception when loading it. Since this file was always
loaded before PC attempted to clean up it could never be deleted since
it was always locked. Not re-enabling the package after a failed attempt
solves this. Note that the package is later automatically removed from
the ignored_packages because of in_progress_packages.

Actually fixes #835.