Re-enable submitting usage info about Package Control being installed
Previously this happened via PackageCleanup via a mechanism designed
to handle upgrades from old versions of Package Control. Since the
3.0 release there has been a huge drop in install stats for Package
Control because we are not explicitly tracking it.

The way it is being submitted should allow for the SSL fixes to be
applied to Linux and Windows before it tries to use the SSL
connection. There may need to be some more tweaks to the ssl-linux
and ssl-windows loaders to get this to work reliably.