Refactor message formatting in source code
There were all sorts of horrible sections of code that were
responsible for generating multiline messages to display to the
user and print to the console when in debug mode.

This cleans those up with a consistent style and some helper
functions that allow for consistent use of multiline indented
strings in source code that are automatically formatted for
user-friendly output.

This gets ride of lots of line continuations, embedded newline
escape sequences and more.