Refactor package listing into a single method
There was lots of duplicated code and duplicated operations on the file
system which a) slowed down the code (even though not a lot) and b) was
just somewhat ugly.

Integrity verified with tests.
Through all the previously used methods and their new equivalents, a
total speedup of ~20% was achieved.