Fix installing local dependencies containing loader.py
This fixes #1242 by adjusting the code in package_manager.py to decode
the loaded contents of loader.py or loader.code to utf-8 prior to
returning it back to the caller.

This mimics what happens in loader.py in the _existing_info function,
which also potentially returns the loader code after loading it in the
same manner and doing a decode to utf-8 prior to return as well.