Avoid unrequired cache exist check in caching downloader
Remove the `cache.has()` query and try to `cache.get()` the key directly as the later one will return False if the cache file does not exist without raising an exception. The same strategy is already used in `cache_result()`. This reduces file usage as an `os.path.exists()` call opens the file to check its existence and closes it (at least on Windows).