Decouple 'tests' package from 'tests_command' module
The more tests are added the more cross references between the `tests` package and the `package_control_tests_command` module are created, while the `tests` package should keep a self containing package. It is important as tests might not need to be part of the public `sublime-package` in the future. They might be excluded via .gitattributes some day.

The package_control_tests_command's only task is to provide an entry point to call `tests.run_all()`.

To avoid recursive import loops, the constants are moved from `tests.__init__.py` to `tests.consts.py`.