Fix package creation destination not created
This commit fixes an issue which caused package_manager to create the
"installed_packages_path" instead of `package_destination`. As a result
ZipFile may fail to create the archive due to missing destination
directory, in case a package is to be created at a user defined
location by `CreatePackageCommand`.

ZipFile(..., "w", ...) makes sure to truncate existing archives, hence
it is not required to explicitly remove an existing archive before.