Optimize 'package_cleanup' module
This commit introduces the follwing changes

1. Optimizations for `clean_old_files()`

It is called as as late as possible because package-control-old files were most likely created by a former `clear_directory()` call, which created a `package-control.cleanup` or `package-control.reinstall` file. Both result in the directory to be cleared out anyway. Looking for old files first is useless then. The `package_manager.install()` method is the only one, which might leave old files behind.

Furthermore walking through vcs based packages can take a very long time because of the huge number of database files. The cleanup-run can easily take several seconds or even 10th of seconds depending on storage speed and package structure, even though the `.git` or `.hg` files won't most likely contain old files. Therefore `clear_directory()` creates a `package-control.old` file in the root directory to mark it for cleanup of old files. If that file doesn't exist `clean_old_files()` ignores the folder.

2. Optimize `clear_directory()`

The advantage of `os.walk` is the ability to loop through directories and files separately. Therefore it is a bit tedious to merge the lists and call `os.isdir()` again to distinguish files and directories. The function is changed to make use of the separate lists to avoid any further filesystem usage as os.walk() already did for us.

3. Use `clear_directory()` for `delete_directory`

shutil.rmtree is a nice function, but finally doesn't handle things much different, than `clear_directory()`. As the later one already implements a strategy to clear a directory, it can be used to prepare directory deletion.