Tweak swapping of 0_package_control_loader
Previously we would potentially be making modifications to the file
0_package_control_loader.sublime-package right around the file we would
tell Sublime Text to take it out of the ignored packages list. This
seemed to sometimes cause a BadZipFile exception about an invalid
header.

Now we ensure that at least 900ms has elapsed since we last made any
changes, in an effort to allow things to be flushed to the filesystem.
Thus results in less churn with ignoring and re-enabling the
0_package_control_loader and will hopefully result in it not getting
stuck in limbo.