Add QuickPanelItem usage to ExistingPackagesCommand
This commit modifies the ExistingPackagesCommand to also use
QuickPanelItems as it is not very obvious which command uses which kind
of quick panel item, which may lead to confusion and bugs.

This change fixes a regression introduced with 3287efd5c1a744b9b8079e5d16aaebf14a2d43a7,
which caused "Package Control: Remove Package" to break, because of the
assumption it would also use the new style of QuickPanelItems.