Fix GitHubUserProvider/GitLabUserProvider
The `failed_sources` attribute of any provider is designed to be a dict.
The `cache['get_packages']` attribute takes packages only.

Hence it appears both GitHubUserProvider & GitLabUserProvider assign
wrong values in case fetching user_info() fails, which is corrected by
this commit.

Furthermore a single `raise` should be enough to re-raise the exception
to the caller.