Fix missing cache_length in DownloadManager
This commit adds `cache_length` to internal settings as this key is queried every time a downloader calls `set_cache()` functions.

The missing value results in cache module always applying a TTL of 300s
instead of the value from Package Control.sublime-settings.

This results in HTTP queries after 5mins, which all result in 304 with
cached values from HttpCache being used and reparsed. This should happen
only if ST was closed.