Fix cli_downloader on Windows
This commit enables usage of cli_downloader on Windows OS by fixing 2
issues.

First of them prevented `curl.exe` or `wget.exe` from being found by
`find_binary()` method as it missed to add the file extension.

Second fix addresses an issue with console window being displayed if
an executable is invoked, which caused bad user experience.

curl.exe is shipped by default as of Windows 10 and therefore might be
used as downloader. May be quite uncommon as the others do a good job,
but if someone decides to use it, it should work.