Add/Remove `installed_packages` directly in thread
As ST3's API is thread-safe reading/writing settings files doesn't need
to be delegated to main thread via `set_timeout()` anymore.

Also those extra checks with regards to `original_names` don't seem to
make sense and are removed. The presence or absence of removed or 
installed packages is checked in the modified methods so there is no
need to pass those `original_names` for the same purpose to
`save_list_setting()` which are now called only if items are added or
removed to the list.