Fix Version Selector (#1579)
* Fix Version Selector

Fixes #1578

A range selector is an greater than or equal and smaller than or equal
operation, which includes provided min/max boundaries into valid range.

It turned out `is_compatible_version()` handles `>=` vs. `>`
and `<=` vs. `<` wrong in general.

* Fix Version Selector

This commit makes `is_compatible_version()` testable and adds some
unittests to validate changes.

Main change is to implement the function in a more explicit way which
makes it harder to fail.

* Add VersionSelectorTests to new TEST_CLASSES list