Fix WinINetDownloader basic authentication
This commit explicitly adds basic auth to request header instead of
passing `username` and `password` to wininet.InternetConnectW().

The reason is simple: The latter doesn't work. Passing username and password seems to have no effect on Windows 11.

As a result github limits API calls to 60 per hour.
With this change 5000 are applicable.

That's an important change in case a custom repository is used rather
than packagecontrol.io channel.