Reduce code duplicates with regards to readme url extraction
This commit moves GithubClient's and GitlabClient's readme url
creation to `_extract_repo_info()` method.

It follows the design of BitBucketClient.

The goal is to avoid duplicated code for URL creation as it was a
reason for a recently fixed bug.

It also makes repo_info() and user_info() methods smaller.