Read all available releases from code hoster
addresses: https://github.com/wbond/packagecontrol.io/issues/143

This commit introduces a generator to fetch all tags until desired
amount of releases is found.

Before this commit clients downloaded first 100 tags to fetch
releases. If a repository however contains many releases with
version-prefixes, it is likely for some of them to be ignored due to
this restriction.