Keep downloaded zip files in memory
This commit uses io.BytesIO to keep a downloaded zip file in-memory,
instead of streaming it to a temporary file on disk.

Data already exists in RAM as it is downloaded as a whole. Just
wrapping it into a BytesIO object so ZipFile can handle it.

Maybe a bit faster and reduces risk of left-overs as all methods
seem to prematurely exit without proper cleanup in error situations,
causing even risk of keeping file handles open!