Fetch packages/libraries once for lifetime of PackageManager
A new PackageManager object is instantiated for each task. The list of
available packages and libraries is very unlikely to change in that
period. Therefore it is sufficient to build it once, only.

So any subsequent call of list_available_packages() and
list_available_libraries() becomes very cheap, which may be important
on batch tasks such as automatic_updater, install_missing, ... which
can easily install/upgrade a large amount of packages.