Replace ThreadProgress by ActivityIndicator
This commit implements a new ActivityIndicator context manager, which
is used to wrap async code to indicate its activity and progress.

It is more flexible with regards how background tasks are scheduled
and executed than relying on a single thread being monitored.

It requires and enables some threading related structural changes of
how tasks are executed.