Refactor ExistingPackagesCommand and ancestors
This commit drops threaded listing of existing packages as creating
lists of present packages is fast but creating threads is rather slow
and expensive.

It's achieved by implementing package list creation and selection
directly in ExistingPackageCommand, which is the new base class for

  - ListPackagesCommand
  - ListUnmanagedPackagesCommand
  - RemovePackagesCommand

As a result quick panel is displayed immediately in all three commands.

Another change prevents permanent PackageManager instances in
WindowCommands as it is designed to live not longer than a task is
running. Otherwise to many or even out-dated cached values may be kept
in RAM.

Same applies for the `self.package_list` members of commands which
display quick panels.