Refactor Enable/DisablePackageCommand
This commit refactors Enable/DisablePackageCommand to follow
implementation of List/RemovePackages.

By inheriting ExistingPackagesCommand look and feel is exactly the
same for all commands which handle locally present packages.

Implementation details of rendering and selecting quick panel items are
isolated in the base class. No need for all the commands to bother with
presence or absence of QuickPanelItem support.