Reorg PackageInstaller ancestors
Just get some weird things sorted and try to introduce a layer concept.

It's pointless for PackageInstaller to implement an `on_done` which is
used only once in InstallPackageThread.

Hence restructure code to implement `on_done` in the front-end threads
InstallPackageThread and UpgradePackageThread with both being based on
PackageInstaller as back-end helper/task which provides required common
functionality to get the job done.