Drop view specific color scheme in case of errors
View specific color schemes are removed if they are provided by a
package, which is going to be upgraded/restored.

Therefore the global color scheme applies during upgrade or
permanently after failure.

It was not a good UX to assign and keep Mariana to views, if global
color scheme is a different one.

Whether Mariana is set, depends only on whether the global color
scheme's package being upgraded.