Refactor PackageCleanup orphaned package removal
A package may consist of both a .sublime-package file and a correspon-
ding overriding unpacked package directory. The packages-metadata.json
most likely exists in one of them only.

Therefore both must be deleted at the same time. Removing the sublime-package file which contains the packages-metadata.json first,
leaves the overriding package directory behind as un-managed.

To avoid that, both must be removed at the same time.

Furthermore orphaned packages are very likely not to be ignored already
as there's no indication for an aborted removal. So we might want to
first determine a list of all packages, disable and remove them in a
batch process to avoid unnecessarily often changes of `ignored_packages`.

As those packages may already contain locked files, removal may need to
be deferred to next start.