Refactor library cleanup
Orphaned libraries are to be removed after all orphaned packages have been removed and all missing packages are installed.

Otherwise it's likely to happen to remove a package, which is directly
re-installed a second later.

As such it seems not very useful to distinguish PackageCleanup and
AutomaticUpdater. At least the step of installing missing packages
doesn't seem to belong to the latter.