Don't block package install/upgrade due to missing libs
The dependencies.json doesn't make any statement about a library being
optional or crucial.

Example:

   UnitTesting requests 'coverage' library, but it's optional. Tests
   work pretty well without it.

   GitGutter does the same with various libraries.

As long as there is no way to distinguish crucial and optional
libraries, we should not block install/upgrade of packages due to
unavailable libraries.

Same applies for libraries which can't be updated in-place, but need
a restart of ST to complete.

It's within the scope of a package to handle missing libraries, maybe
by prevent starting/loading sub-modules.