Change current working directory only to remove a directory
Fixes #1303

All functions work with absolute paths. So there's no need to change
current working directory to extract zip files.

To successfully remove a real directory however cwd must not point to
one of its sub directories. Otherwise PermissionDeniedErroe with
WinError 23 is raised. It doesn't apply to junctions or symlinks.

For simplicity reasons cwd is set to parent dir in clear_directory().
It makes sure both - a directory can be cleared and deleted.