Improve Error handling of sys_path.cache_path()
It turned out sys_path functions returning `None` may cause a lot of
trouble or needs special treatement to avoid bad things from happening.
At least type checkers such as pyright mark lots of source as error
prone.

As it is expected sublime.cache_path() to return valid strings normally, but fail only on ST3 when API is not yet ready, it seems
more sufficient to raise a RuntimeError if it returns None, to stop
all sub-sequent execution of code.