Split Advanced Install Packages Command
Current `install_packages` command does both, it installs packages,
which don't exist or upgrades/overwrites them otherwise, regardless
whether they are up-to-date or not. The latter may not be expected.

This commit restricts `install_packages` to only install specified
packages, if they don't exist. A new `upgrade_packages` command is
added which upgrades existing packages, but ignores those which are
not installed.