Refactor Install/Upgrade command classes
Drop separate Thread classes, which are also PackageInstallers.

Use closure functions instead, to handle life cycle of required objects
to comply with Python's "as local as possible" style guide.

It also seems a bit odd to create a `Thread` for a task, while running
subsequent async operations via `threading.Thread(target=...)`.

The goal is to run async tasks in a thread pool using a global
`ThreadPoolExecutor` object in the future.