Fix sys_path ZipLoader detection
This commit fixes an issue which caused package_control.sys_path to
detect correct paths, in case Package Control is loaded from
*.sublime-package file.

The reason is the `__loader__` being a `sublime_plugin.ZipLoader`
instance rather than the `ziploader`, which was checked against before.

As a result the module returned wrong paths, causing the bootstrapper
to fail converting dependencies to libraries.

Notes:

1. The major reason for that song and dance is ST3 not returning valid
   values when calling sublime.packages_path() before API is ready.
2. Second reason seems compatibility with ST's dev environment at
   sublimehq.
3. The order of code branches is swapped as it is much more likely to
   load Package Control via ZipLoader.