Reduce API calls when fetching repositories
This commit adds an internal `min_api_calls` setting, which is set true
when running in ST to only fetch least required information from code
hosters. The goal is to reduce API calls and hit rate limits later.

Fetching `timestamp` for each tag is too expensive to just drop the
returned information when running in ST.