Teach bootstrapper to mark dependencies for cleanup
This commit treats converted dependencies as normal packages with
regards to removal after conversion to libraries. In case a dependency
folder can't be removed, it is marked for cleanup at next ST start.

Using `delete_directory` makes sure to treat symlinks and junctions
correctly on Windows OS, in case a dependency contains one.