Refactor Bootstrapping
This commit moves bootstrapping code to `package_control` and loads it within
main 'Package Control' plugin, so bootstrapping is not performed before it is
clear, all requirements for running Package Control are met.

It also ensures, Package Cleanup to be executed after bootstrapping is done
without the need of setting/reading `bootstrapped` setting.