Main Package Control plugin
This commit merges `1_reloader` and `Package Control` modules into a single
`plugin.py` as it turned out to be the most robust architecture and is
therefore already common practice in various other complex packages.

As of ST3+ plugins don't need to have unique names and thus it is safe to use
a common name for an main entry plugin. The name `plugin` is used as console
then displays:

    reloading plugin Package Control.plugin