Fix index_files being set False on Package Control updates
This commit persists old user value of `index_files` settings on disk as
global variables/states are vanished during Package Control updates, causing
`index_files` not being reset afterwards.

It uses json serialization to be able to extend backup usage in future,
without loosing backward compatibility. As Package Control blocks all other
updates, if it is going to update itself, it is very unlikely other settings
stored by PackageDisabler need to be persisted for the purpose of surviving
Package Control update.

There may however be other use cases to do so, which are not within scope 
of this fix.