Add support for python specific legacy dependencies
This commit extends support for legacy dependencies format.

Specifically dependencies may now also contain following folders for
ST4 / python specific code paths:

  - "st4"                   = install only on ST4
  - "st4_{PY}"              = ... and given python version
  - "st4_{PY}_{OS}"         = ... and given OS
  - "st4_{PY}_{OS}_{ARCH}"  = ... and given architecture

If both "st3" and "st4" folders are present, the latter one is preferred on ST4
and ignored on ST3.

Platform/arch specific ST3 dependencies are restricted to python 3.3 as those
are assumed to contain compiled libraries.

Support is added to help those more familiar with the old format.

It is however not recommended as the whole library with all variants needs to
be downloaded which costs more bandwidth and may therefore result in slower
installations.

It is recommended to create or re-use existing WHEEL files and add a release
for each python/platform/arch combination.