Fix python specific library version selection
This commit fixes an issue which caused a library from not being installed,
if the most recent version is not available for the requesting python versions.

Specifically, if a library's v2.0.0 is available for python 3.8+, while only
v1.0.0 is available for python 3.3, installation/upgrade failed as PC wanted
to install v2.0.0 on python 3.3 as well.