Optimize http cache ttl
This commit updates file times, when reading cache object from disk, so TTL
applies only if a file is no longer accessed for a certain period of time.

It's nonsense to delete a cache file, just because it has not been modified for
a certain period of time. Actually most cache files of github API calls remain
unchanged for years as packages are not updated.