Drop library dependencies
Dependencies are installed as python libraries, which causes ST package style
dependency information to get lost. As a result subsequent calls to
"satisfy libraries" wouldn't be able to restore required dependency information
and thus remove those libraries again.

As PC3.x didn't support ST dependencies with dependencies, it's most safe to
restore that state for the time being as converting them to "requirements.txt"
may be cumbersome and likely to fail as they can't be distinguished from normal
python packages then.

On the other hand maintaining 3 independent dependency systems for packages,
legacy dependencies and python packages is probably not worth it.

2 are enough for a future plan.