Fix ignored_packages setting getting changed at startup if default packages have been extracted
If the default packages aren't in a `.sublime-package` file then on
startup the `"ignored_packages"` setting would always be set to `[]`.

Primarily this fixes our development environment where we don't use any
`.sublime-package` files, but presumably there are also people who've
extracted these folders (as unwise as that may be).