Fix vcs package upgrade task termination
This commit fixes an issue, which caused an upgrade task not being terminated
in case an invoked git/hg subprocess got stuck. An error message was displayed,
but corresponding `proc.communicate()` never exited.

To fix it, `proc.communicate(..., timeout=60)` is used to gracefully return
in case of stuck sub processes.