Teach `http_get` to read "file:///" urls
We can already host local repositories -- i.e. "repositories" in the
PC settings understand absolute local paths as additional repositories.
But then inside such a repository, we cannot point a release url to a
local file but MUST use the http-protocol.

That seems unnecessary as the "file:///"-scheme denotes a local file
expressed as an url, and frankly, other fetchers would accept them just
as-is.  E.g. a simple `urllib.request.urlopen(location)` would work with
both http and file urls.

Here, it doesn't work because of basic tests in `_grab`, e.g.:

```
        parsed = urlparse(url)
        if not parsed or not parsed.hostname:
            raise DownloaderException('The URL "%s" is malformed' % url)
```

Don't scratch these tests but instead check for the "file:///" scheme at
the top, and do a straight, dead-simple `file.open` instead.

That's the minimal patch to support the usecase.