Host main channel.json as GitHub release asset
To get hosting "for-free" and to enable ETag and Last-Modified support,
cache the official "channel_v3.json" on GitHub.  Ensure to update the
release asset only when the contents of the channel actually changed.

It is possible to post-process the "channel.json" in the `generate`
phase, i.e. we could drop old ST2 only packages.  But that is not part
of this PR.