Install further packages automatically
PC has the notion of "dependent" or additional packages.  These are
defined in a standard "Package Control.sublime-settings" file, hosted
by the package.  (I.e. similar to the "dependencies.json".)

Installing such a package did not install the additional packages in one
go, neither did it notify about the incomplete install.  In the end, a
restart is enough and required to trigger `install_missing_packages()`
to catch up and install everything.

We have two options here.  (1) to bail out `install_package` with
`return None` and a message to tell the user to restart Sublime, or (2)
to install the packages in a recursive style.

Here we choose (1) which is a simple babystep for a better UX.  (2) is
not possible as all install side-effects need to run serialized.