Do not cache "file:///" resources (#1697)
Fixes #1695

We generally cache to save the network latency, not because our
transformations of the data were so expensive.

We expect that "file:///" protocol resources can be read fast enough for our
purposes and want to ensure that changes of such local files are available
immediately, e.g. without restart Sublime Text or clearing the cache as a
whole (`clear_cache`).