Fetch unavailable libraries directly from pypi
If a requested library is not available in one of the repositories,
usually the channels "libraries_cache", fetch the information directly
from PyPi.

The core-feature is to automatically discover suitable versions of a
library, even when the Python host changes or gets compiled differently.

Vendors [packaging](https://github.com/pypa/packaging).